// src/app/shared/models/carSummary.model.ts
export interface SelectedModel {
    selectedCode: string;
    selectedColor: string;
    selectedColorPrice: number;
    selectedConfigDescription: string;
    selectedImageUrl: string
    selectedModelDescription: string;
    selectedModelPrice: number;
    selectedRange: number;
    selectedSpeed: number;
    selectedTowHitch: number;
    selectedYoke: number;
}